Responding to 'High' security scan issue
Severity: High
Issue: connect
The second argument specifiying the packed address to bind to, should not be derived from user input. If the address is derived from user input, it is possible for a malicious user to cause the socket to connect to an arbitrary remote address, enabling hijacking of potentially sensitive network data.

    File: 20250701.220714.17650/mlspgs-scripts/l1pcffrag_generator.pl
    Lines: 133